+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [v0.7.2] - (2026 June ??)
 
 - Fixed increment summary flags stollen
+- Fixed UI
 
 ## [v0.7.1] - (2026 June 08)