+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed increment summary flags stollen
 - Fixed UI
+- Added auto font sizing for team names
 
 ## [v0.7.1] - (2026 June 08)